vec.h (VEC_T_length, [...]): Use unsigned, not size_t.
	* vec.h (VEC_T_length, VEC_T_index, VEC_T_iterate, VEC_T_truncate,
	VEC_T_replace, VEC_T_quick_insert, VEC_T_safe_insert,
	VEC_T_ordered_remove, VEC_T_unordered_remove): Use unsigned, not
	size_t.
	(struct VEC): Use unsigned for num and alloc.
	* vec.c (struct vec_prefix): Likewise.
	(vec_o_reserve): Adjust.

From-SVN: r84973

+++ gcc/vec.c
@@ -29,8 +29,8 @@ Software Foundation, 59 Temple Place - Suite 330, Boston, MA
 
 struct vec_prefix 
 {
-  size_t num;
-  size_t alloc;
+  unsigned num;
+  unsigned alloc;
   void *vec[1];
 };
 
@@ -57,7 +57,7 @@ vec_o_reserve (void *vec, int reserve, size_t vec_offset, size_t elt_size
 	       MEM_STAT_DECL)
 {
   struct vec_prefix *pfx = vec;
-  size_t alloc = pfx ? pfx->num : 0;
+  unsigned alloc = pfx ? pfx->num : 0;
 
   if (reserve >= 0)
     alloc += reserve;
